consolidate file explorer feature and notes
